Version 1.6 initial pr (#1370)
* update installation guide & readme (+6 squashed commits)
Squashed commits:
[41634d25] update version numbers
[a10f9b9f] allow tf 2.11.0 or 2.11.1
[c9538bdf] remove erroneously flagged alocc patch, real root cause was problem with BatchNorm2D on mac which is fixed in torch nightly branch
[41d1ca8e] update benchmark
[f05524b2] update system to default to legacy optimizers which appear to be faster on both mac and linux
[3f97f7be] updates regarding optimizer (+19 squashed commits)
Squashed commits:
[2c9fcb03] torch2 only_inputs defaults True
[a283f835] typing fixes
[5e555689] fix alloc torch apphub
[cbc5fc06] fix cyclegan notebook to not manually set device
[da32af29] fix torch unnecessary device passing
[c8bcd350] fix repeated key printing warnings
[e5316e22] rand aug verified
[4a1717d4] formatting
[f8e33c2b] fix type hint
[1e51b7e6] improve typing
[b8554689] updated fgsm
[b8e03612] typing and import fixes
[06e74467] update hinge notebook
[9cd2fd80] update ecc notebook
[7d3daee9] enhance Suppressor class to allow targeted suppression of TF backend messages
[a054a253] allow suppressor to show prints retro-actively on error
[f54ecb25] more unit test fixes
[d9ac7c87] initial unit test work
[f934d4d5] typing fixes

* unused imports

* f-string codacy flag

* set typing version

* update mixed precision optimizer handling

+++ fastestimator/op/numpyop/univariate/onehot.py
@@ -68,6 +68,6 @@ class Onehot(NumpyOp):
 
         if self.label_smoothing != 0:
             smooth_label = self.label_smoothing / self.num_classes
-            output = np.where(output != 0, 1.0 - self.label_smoothing + smooth_label, smooth_label)
+            output = np.where(output != 0, 1.0 - self.label_smoothing + smooth_label, smooth_label).astype(np.float32)
 
         return output
